+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## Version 0.1.0
 
 * Initial commit of required files for a public repo.
+
+## Version 0.2.0
+
+* Initial implementation for Exercise 03 - Connecting to Confluent Cloud.
+
+## Version 0.3.0
+
+* Initial implementation for Exercise 07 - Querying Flink Tables